[#1] fix: statistics.service.ts에서 post레포지토리를 가져오는 것을 postLib를 통해 연결함.
1. statisticsService > postLib > postService > postRepository 흐름으로 연결
